requirements.txt
torch
torchvision
logging
Pillow
numpy
scipy
scikit-learn
tqdm